Cluster와 ClusterCoordinator 병합
